requirements.txt
email-validator>=2.2.0
flask>=3.0.3
flask-sqlalchemy>=3.1.1
psycopg2-binary>=2.9.9
flask-jwt-extended>=4.6.0
flask-login>=0.6.3
flask-migrate>=4.0.7
flask-wtf>=1.1.1
wtforms>=3.1.1
python-dotenv>=1.0.0
pytz>=2024.2
waitress>=2.2.1
